TST: implement xfailed test for #18415
